address review comment: merge into reference_utils
