+++ CMakeLists.txt
@@ -168,8 +168,7 @@ set(EXTENSION_SOURCES
     src/expression/aggregate_id.cpp
     src/expression/ast/clone.cpp
     src/expression/ast/node.cpp
-    src/expression/ast/substitute.cpp
-    src/expression/ast/visit.cpp
+    src/expression/ast/reference_utils.cpp
     src/expression/from_duckdb.cpp
     src/expression/function_id.cpp
     src/expression/join_condition.cpp
