+++ src/planner/sirius_plan_projection_utils.cpp
@@ -16,8 +16,7 @@
 
 #include "planner/sirius_plan_projection_utils.hpp"
 
-#include "expression/ast/substitute.hpp"
-#include "expression/ast/visit.hpp"
+#include "expression/ast/reference_utils.hpp"
 #include "op/sirius_physical_projection.hpp"
 
 #include <cstddef>
